refactor: decouple server health helpers
